+++ examples/04-mock-express/demo/project/data/index.js
@@ -1,4 +1,0 @@
-
-module.exports = (req, res) => {
-    res.json({success: !0, message: 'index'});
-};