+++ examples/04-mock-express/index-port.js
@@ -1,7 +1,0 @@
-
-var projectConfig = require('./demo/lila.config');
-projectConfig.mockExpressServerPort = 8191;
-
-require('../../util/change_cwd_to')(__dirname + '/demo');
-
-require('../../util/exec')('lila mock-express');